1. fix a bug in binary_dataset 2. move window dataset to a new file
