+++ time_moe/datasets/time_moe_dataset.py
@@ -9,10 +9,13 @@ from .binary_dataset import BinaryDataset
 
 
 class TimeMoEDataset(TimeSeriesDataset):
+
     def __init__(self, data_folder, normalization_method=None):
         self.data_folder = data_folder
         self.normalization_method = normalization_method
         self.datasets = []
+        self.num_tokens = None
+
         if normalization_method is None:
             self.normalization_method = None
         elif isinstance(normalization_method, str):
@@ -83,61 +86,11 @@ class TimeMoEDataset(TimeSeriesDataset):
         dataset_offset = seq_idx - self.cumsum_lengths[dataset_idx]
         return self.datasets[dataset_idx].get_sequence_length_by_idx(dataset_offset)
 
+    def get_num_tokens(self):
+        if self.num_tokens is None:
+            self.num_tokens = sum([ds.get_num_tokens() for ds in self.datasets])
 
-class WindowDataset:
-    def __init__(self, dataset: TimeSeriesDataset, context_length: int, prediction_length: int = 0):
-        self.dataset = dataset
-        self.context_length = context_length
-        self.prediction_length = prediction_length
-        self.window_size = context_length + prediction_length
-
-        self.window_info_list = []
-        n_seqs = len(self.dataset)
-
-        cur_window_info = []
-        num_cur_remaining_points = self.window_size
-        for seq_idx in range(n_seqs):
-            seq_len = self.dataset.get_sequence_length_by_idx(seq_idx)
-            remaining_seq_len = seq_len
-            while remaining_seq_len > 0:
-                if remaining_seq_len < num_cur_remaining_points:
-                    cur_window_info.append(
-                        (seq_idx, seq_len - remaining_seq_len, remaining_seq_len)
-                    )
-
-                    # update states
-                    num_cur_remaining_points -= remaining_seq_len
-                    remaining_seq_len = 0
-                else:
-                    # add the part of this seq to cur_window
-                    cur_window_info.append(
-                        (seq_idx, seq_len - remaining_seq_len, num_cur_remaining_points)
-                    )
-
-                    # update states
-                    remaining_seq_len -= num_cur_remaining_points
-                    self.window_info_list.append(cur_window_info)
-
-                    # reset current window
-                    num_cur_remaining_points = self.window_size
-                    cur_window_info = []
-
-        if num_cur_remaining_points > 0:
-            # TODO drop last temporarily
-            pass
-
-    def __len__(self):
-        return len(self.window_info_list)
-
-    def __getitem__(self, window_idx):
-        window_info = self.window_info_list[window_idx]
-        seq = []
-        for seq_idx, start_idx_in_seq, offset  in window_info:
-            part_seq = self.dataset[seq_idx][start_idx_in_seq: start_idx_in_seq + offset]
-            seq.append(part_seq)
-        seq = np.concatenate([seq], axis=0)
-        return seq
-
+        return self.num_tokens
 
 
 def zero_scaler(seq):
