+++ time_moe/datasets/ts_dataset.py
@@ -12,10 +12,9 @@ class TimeSeriesDataset:
     def __getitem__(self, seq_idx):
         pass
 
-    def __iter__(self):
-        n_seqs = len(self)
-        for i in range(n_seqs):
-            yield self[i]
+    @abstractmethod
+    def get_num_tokens(self):
+        pass
 
     @abstractmethod
     def get_sequence_length_by_idx(self, seq_idx):
@@ -23,4 +22,9 @@ class TimeSeriesDataset:
 
     @staticmethod
     def is_valid_path(data_path):
-        pass
+        return True
+
+    def __iter__(self):
+        n_seqs = len(self)
+        for i in range(n_seqs):
+            yield self[i]
